wrap enumTest in IIFE
the `var expect` declaration is hoisted, but because the script is
executed at top-level scope, it is wiping/undefining any previously set
value of `expect`. So the boolean OR then attempts to assign via a
require statement which fails, of course.

Solution is to ensure the `var expect` is scoped within a function so
that it doesn't wipe any existing value of expect on `global` (`window`)